Refactor widget location and implementation
